Pipewire buffer size controller
I thought I would make a small app for changing the buffer size of pipewire which somes in handy for music production. Currently I have had keyboard shortcuts to execute shell commands for the desired buffer size but it seems like a good idea for a first GUI project aswell.