Changed error handling
I gave up on the error button. Why it didn't work is beyond my capabilities at this stage.

Instead I spent time on doing a check if pipewire is even installed and making different errors depending on if it is or not which seems like a waste of time since no one will probably ever use this app except for me!