added label for settings
Added labels for settings. Upon start it takes the output of a shell command to list pipewire settings. I the format it to a list and select the indexes in the labels. I hope pw don't change the format of their current settings page, If they do I guess I will use the setting name and append index+1 to a new list and use those, or something like that.

I also added a light and dark mode GUI window example for anyone who might want to see the result but not running linux with pipewire.

Lastly I cleaned up a little bit in the glade-file and changed labels and names of things to better reflect what is happening as well as cleaning up some commented away test lines I used for troubleshooting before