Separating business and GUI logic
Separated the logic better between the classes and changed the method to get currents settings slightly.

Next part is to figure out why every line in ControlWindow needs self. infront of it to work EXCEPT in the on_error() method and see if I change further change get_current_settings() to not break from potential formating differences in pipewire cli