Fixed get_current_settings()
I found a bug in get_current_settings where they both showed '0' after a reboot which was because the output I used for it was the currently FORCED setting and after a reboot there is no forced and it uses a DEFAULT setting instead. I also changed the way of getting the output to something I hope if less prone to breaking if pipewire changes the way it lists it's settings

Lastly. Added some comments with questions I have for anyone seeing this that would be willing to help