Finally "works"
It's now properly separated into classes for handling the GUI and handling the logic

I have Fixed formatting in the output so it's the same as the input

I addressed special characters and make sure to convert (the ones I have come across) into decimal form so they can be multiplied.

Fixing a slight quirk that makes the program add .0 decimals to some conversions. for example 50% of 300 g would be 150.0 g. But no longer

I should add that I have not yet tested with a lot of different inputs so I don't know of all the cases that it could be tripped up. After, free user input has a lot of potential to be parsed unexpectedly :D