Fixed GUI resize
I fixed a few things, for example adding a scrollbar then the text can't fit in the input and output boxes.

I Also made it so that the boxes are set to fill the size of the window In whatever size it is in but requesting a 300 points height as default

I adjusted the packing of the close button to accommodate for the re-sizable window by putting an empty column between it and apply that also expands to fill the window horizontally which forces the close button to the right bottom corner at all times instead of floating somewhere randomly! :D