Changing GUI
Experimented a bit in Glade and realized that the GUI i made had huge problems with scaling which is why I just had the windows size fixed.

In the light of this I decided to try to redesign it in a different way that is more robust to different sizes.

I have a first box with 2 sections:
1. all the setttings
2. both buttons and an empty middle that expands horizontally so that the close buttons stays in the right corner

In the top section I have a grid with 1 row and 11 columns:
1 separator
2 sample size
3 margin label*
4.separator
5. margin label*
6. Sample rate
7. margin label*
8. separator
9 margin label*
10 current settings
11 separator

Margin label = empty label that expands with window horizontally.
I assume that there are better way to achieve the same effect but This made sense to me.

Also changed out example photos to the new GUI